fix: wrong queueLength count with tempQueue
also: split storage into several files for better maintainability
