+++ src/options/options.ts
@@ -2,13 +2,14 @@ import { storage } from "webextension-polyfill";
 import Badge from "../Badge";
 import { injectIcons } from "../icons";
 import { localizeUI } from "../Localization";
-import Storage, { Key } from "../Storage";
+import OptionsStorage from "../storage/OptionsStorage";
+import QueueStorage from "../storage/QueueStorage";
 import AdBlockSwitcher from "./AdBlockSwitcher";
+import AnimationLevelSlider from "./AnimationLevelSlider";
 import BlockSpeedSlider from "./BlockSpeedSlider";
 import ImportExport from "./ImportExport";
 import "./options.scss";
 import ScrollSpeedSlider from "./ScrollSpeedSlider";
-import AnimationLevelSlider from "./AnimationLevelSlider";
 
 function initOptionSettings() {
 	new AdBlockSwitcher();
@@ -28,10 +29,7 @@ function initOptionSettings() {
 
 	if (resetButton) {
 		resetButton.addEventListener("click", async () => {
-			await Storage.remove(Key.blocksPerMinute, false);
-			await Storage.remove(Key.scrollsPerMinute, false);
-			await Storage.remove(Key.animationLevel, false);
-			await Storage.remove(Key.adBlockerActive, false);
+			await OptionsStorage.resetSettings();
 			initOptionSettings();
 		});
 	}
@@ -40,7 +38,7 @@ function initOptionSettings() {
 		clearButton.addEventListener("click", async () => {
 			await storage.local.clear();
 			initOptionSettings();
-			const queueLength = await Storage.getQueueLength();
+			const queueLength = await QueueStorage.getQueueLength();
 			Badge.updateBadgeCount(queueLength);
 		});
 	}
