+++ src/settings.ts
@@ -16,5 +16,6 @@ export default {
 		normal: "https://abs.twimg.com/sticky/default_profile_images/default_profile_normal.png",
 		mini: "https://abs.twimg.com/sticky/default_profile_images/default_profile_mini.png",
 	},
-	DEQUEUE_BATCH_SIZE: 50,
+	DEQUEUE_BATCH_SIZE: 42,
+	TEMP_QUEUE_BATCH_THRESHOLD: 14,
 };
